Hoist location and dependencies to common module
