+++ app/src/main/java/io/homeassistant/companion/android/sensors/SensorReceiver.kt
@@ -12,6 +12,7 @@ import io.homeassistant.companion.android.BuildConfig
 import io.homeassistant.companion.android.common.sensors.BatterySensorManager
 import io.homeassistant.companion.android.common.sensors.LastUpdateManager
 import io.homeassistant.companion.android.common.sensors.SensorManager
+import io.homeassistant.companion.android.common.sensors.NetworkSensorManager
 import io.homeassistant.companion.android.common.sensors.SensorReceiverBase
 
 @AndroidEntryPoint
